Use classic typography scale where h1 equals base font * scale.
This is correcting my deep ignorance of last year on this stuff. This
article helped immensely
http://spencermortensen.com/articles/typographic-scale/

This is a major breaking change as everyone's headers will be
significantly smaller now until they update their scale(s).

As part of this, the new defaults are "octave" (this the traditional
typography scale) and for responsive, "golden" as that's always a golden
choice.