Make styles generation a two-step process
Previous to this change, the css rules were generated by joining
individually generated rules defined in somewhat ad-hoc ways. Now we
first generated a object tree of rules that correspond to a selector and
rule set and then recursively compile this into a styles string.

This change paves the way for themes to be able to modify the styles
before compilation and a plugin system.