Use number ratio for line-height instead of pixels
This is both more useful (can scale font size while keeping line height
same) and more common. Most people just use 1.5 or whatever for
line-height not calculating out the pixel.

I would just move everything over to the number ratio way of doing
things but right now compass-vertical-rhythm which is doing a lot of the
calculations under the hood expects baseLineHeight to be a pixel value
so we're stuck for now.