Fix size of <pre> + give code blocks own line-height
It doesn't really make sense to let code blocks inherit the line-height
of normal body text because it *isn't* normal body text.

It makes more sense to treat it like an image or some other arbitrary
element dropped in a page and not expect it to maintain the vertical
rhythm.