Adjust default header sizes
Previously h1-h6 each took their size by dropping down one note on the
classic typographical scale. So if base font size was sixteen and the
scale ratio is 2 the h1 was 32, h2==~28, h3==~24, h4==~21, h5==~18,
h6=16.

This is great except... in practice most sites style h1s and h2s with a
greater contrast between them. The header sizes Typography.js was
pumping out made in hard to distinguish sometimes between h1s and h2s
which shouldn't happen. I was actually starting to change my h2s to h3s
when I started looking closely at Github's styles and realized that they
had just skipped a note when styling h1/h2 so h1 was 32 but h2 was 24,
not 28.

This seems like the practical way to set header sizes as while in theory
it might be nice if people were to understand how to use different headers
to get the right contrast between headers or whatever, in practice,
people (including me) are just going to use h1s and h2s and expect them
to look good.