Reduce baseLineHeight to 1.45 + use fixed line height for headers
This is following advice from the Typography Checklist https://www.typewolf.com/checklist

On the baseLineHeight change, they say 1.5 is standard but that less is
probably preferable with 1.45 being a better standard so we'll adopt
that.

Then he complains about people a) using the same line height for headers
for bodies and b) sticking to rigidly to a "baseline grid" for
  typography as choosing the correct line height is more important than
  sticking to a grid.

Typography.js is guilty of both 😩

I've been noticing this for a while, that I always override the line
height for headers as Typography.js is creating too large of gaps.

So as part of this PR I've added a new `headerLineHeight` option and set
its default to 1.1 which seems like a reasonably safe default.